[REFACTOR] #6 - presentation layer 분리
